- implemented OpenGL prefs menu (only flags) - fixed possible reference to freed memory in RenderVisTree::cast_render_ray()   which caused Desla 2 to crash - fixed reference to unallocated memory in possible_intersecting_monsters()
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@491 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/map.h
@@ -622,7 +622,7 @@ struct side_data /* size platform-dependant */
 	int16 secondary_lightsource_index;
 	int16 transparent_lightsource_index;
 
-	int16 ambient_delta;
+	int32 ambient_delta;
 
 	int16 unused[1];
 };
